Elevator and small game design progress
The elevator works almost perfectly, there are problems when the player gets stuck to the side of it, but for now i will leave it for now.  I also made an incredibly rudimentry version of the cave, that is half finished. The camera currently is not adjusted to he cave. Will need to make an object to cause changes in camera setting. If that makes sense.